+++ mobile/src/features/Chat/ui/Chat.tsx
@@ -1,6 +1,6 @@
 import {RouteProp, useRoute} from '@react-navigation/native';
 import React, {useState} from 'react';
-import {FlatList, KeyboardAvoidingView, Text, View} from 'react-native';
+import {KeyboardAvoidingView, View, VirtualizedList} from 'react-native';
 
 import {Screens} from 'app/navigation/navigationEnums';
 import {RootStackParamsList} from 'app/navigation/navigationTypes';
@@ -15,20 +15,19 @@ type ProductRouteProp = RouteProp<RootStackParamsList, Screens.MESSENGER>;
 export const Chat = () => {
   const route = useRoute<ProductRouteProp>();
   const {chatId} = route.params;
-  // TODO: Получать это все из стора useChatStore
+
+  // TODO: Получать это все из useChatStore
+
   // const {messages, currentMessage, loadMessages, sendMessage} = useChatStore()
+
   // TODO: Удалить / поменять
   const [message, setMessage] = useState('');
 
   const chat = chats.find(chat => chat.id === chatId)!;
 
-  const sellerName = `${chat.product.user.name} ${chat.product.user.surname}`;
-
   const messages = chat.messages;
 
   const currentUser = users[0];
-
-  // TODO: Удалить / поменять
 
   const handleSendMessage = () => {
     setMessage('');
@@ -36,9 +35,13 @@ export const Chat = () => {
 
   const handleAttachFile = () => {};
 
-  //---------
+  //-------------------------------------------------------
 
-  const renderMessageItem = ({item}: {item: IMessage}) => {
+  const getItemCount = () => messages.length;
+
+  const getItem = (data: IMessage[], index: number) => data[index];
+
+  const renderMessage = ({item}: {item: IMessage}) => {
     const isCurrentUser = item.authorId === currentUser.id;
     return <Message message={item} isCurrentUser={isCurrentUser} />;
   };
@@ -46,10 +49,13 @@ export const Chat = () => {
   return (
     <View style={styles.container}>
       <KeyboardAvoidingView style={styles.messagesContainer}>
-        <FlatList
+        <VirtualizedList
           data={messages}
+          initialNumToRender={20}
+          renderItem={renderMessage}
           keyExtractor={item => item.id.toString()}
-          renderItem={renderMessageItem}
+          getItemCount={getItemCount}
+          getItem={getItem}
           contentContainerStyle={styles.contentContainer}
         />
         <ChatTextarea
