+++ mobile/src/features/ChatList/index.ts
@@ -1,2 +1,1 @@
-export * from './ui/ChatList';
 export * from './model/store';
