--- mobile/src/features/ChatList/ui/styles.ts
@@ -1,13 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-import {Colors} from 'shared/libs/helpers/colors';
-
-export const ChatListStyles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: Colors.White100,
-  },
-  listContent: {
-    paddingBottom: 10,
-  },
-});
