+++ mobile/src/pages/ChatsList/ui/ChatsList.styles.ts
@@ -2,9 +2,12 @@ import {StyleSheet} from 'react-native';
 
 import {Colors} from 'shared/libs/helpers/colors';
 
-export const ChatsListStyles = StyleSheet.create({
+export const ChatListStyles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: Colors.White100,
+  },
+  listContent: {
+    paddingBottom: 10,
   },
 });
